Move Syntax abstractions into state
Since they seem too minor to warrant an own package
