+++ state/src/index.ts
@@ -3,3 +3,4 @@ export {EditorSelection, SelectionRange} from "./selection"
 export {Change, ChangeDesc, ChangeSet, Mapping, MapMode, ChangedRange} from "./change"
 export {Transaction} from "./transaction"
 export {Slot} from "../../extension/src/extension"
+export {syntax, Syntax, SyntaxRequest} from "./syntax"
